+++ esp-phy/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- `last_calibration_result` to get the result of the last calibration
 
 ### Changed